requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
pydantic
alembic
python-dotenv